+++ InsprWall.py
@@ -13,7 +13,7 @@ import requests
 from Desktop import get_desktop_environment, set_wallpaper
 
 NAME = "InsprWall"
-VERSION = "2"
+VERSION = "3"
 
 if "APPDATA" in os.environ:
     CONFIG_FILE = os.path.join(os.environ["APPDATA"], NAME)
@@ -89,15 +89,19 @@ def main():
             valid_sizes.extend(["[4k]", "[4K]"])
         log("Valid sizes: {sizes}".format(sizes=valid_sizes))
 
+        valid_formats = {f for f in CONFIG.get("Other", "formats").split("+")}
+        log("Valid formats: {formats}".format(formats=valid_formats))
+
         for post in req_reddit.json()["data"]["children"]:
             if any(size in post["data"]["title"] for size in valid_sizes):
                 log('Matched post "{title}"'.format(
                     title=post["data"]["title"]))
-                url = post["data"]["url"]
-                break
-            else:
-                log('Skipped post: "{title}"'.format(
-                    title=post["data"]["title"]))
+                image_format = post["data"]["url"][::-1].split(".")[0][::-1]
+                if image_format in valid_formats:
+                    log('Matched format "{f}"'.format(f=image_format))
+                    url = post["data"]["url"]
+                    break
+            log('Skipped post: "{title}"'.format(title=post["data"]["title"]))
 
         background = requests.get(url)
         CACHE.set(uuid4(), background, expire=cache_seconds)